chore: add left label support
